+++ CHANGELOG.md
@@ -6,6 +6,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 [markdownlint](https://dlaa.me/markdownlint/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.0] - 2021-09-21
+
+### Added in 1.4.0
+
+- Added subcommands for Azure Queue and Azure SQL Database:
+
+  - `read-from-azure-queue-withinfo`
+  - `read-from-azure-queue`
+  - `redo-withinfo-azure-queue`
+  - `write-to-azure-queue`
+
 ## [1.3.10] - 2021-08-12
 
 ### Added in 1.3.10